element: ak-notification-counter
`<ak-notification-counter>` implements the Patternfly 5 “notification counter” pattern, which combines a button-pattern for showing notifications with an ARIA-compatible status indicator for the number of notifications outstanding. This instance of the counter allows clients to customize the icon and the content of the notification label, so that multiple kinds of notifications can be mananged.