element: ak-tooltip
`<ak-tooltip>` implements a tooltip using the native `<dialog>` element as its basis, and exploiting a few tips and tricks from Patternfly’s CSS, so it conforms to the look and feel of the Patternfly React version of the component.

In pushing the limits of what is possible, the Tooltip derives stylistic settings from CSS, including the delay for when to show the tooltip! This allows us to configure it so that `prefers-reduced-motion` can be changed to delay the showing and hiding of the tooltip, as well as dicate that the tooltip blink in and out rather than fade. We check the delays on every re-render, so we will pick up changes even in if the user changes the setting in the middle of a session.

There is a strong, functionally-oriented state machine that controls the hide/show delay, but there is also a large explanatory comment.

The `.wccss` is horribly out-of-date. It was an excellent starting point, but the CSS is highly customized at this point. (I really have to add comments to the WCCSS parser).