This commit introduces a dark-mode controller and code.
## Decisions

### Policy Decisions

Setting the attribute `<html theme="dark">` will always take priority over browser settings.

### Conventions

For each component, there are *at least* two CSS files: One that should be bundled with the component as part of its template, so it is converted during the build process into a CSSResult. The other should be imported and mixed into the text of the global CSS file. To distinguish between the two, a filename with multiple periods in its name, such as `ak-component.root.css`, is assumed to be a native CSS or SCSS file, and will not be converted into a CSSResult.

## Code changes

### Theme controllers

The theme controllers introduce a per-component controller, for those components that have explicit light/dark mode settings, and a global controller, which provides a context for the per-component controllers. The global controller reads the attribute `<html theme="light|dark">`; in the absence of a theme attribute, it will use the media query. (See “Policy Decisions” above.)

In the absence of the context provider, the per-component controllers will use the media query exclusively.

### Ak-Brand

The `types` and `template` files were consolidated in the component, as they removed rather than enhanced the readability of the component. This is mostly a personal judgment. For components with a lot of types or a lot of rendering, it will still make sense to keep them separated.

The CSS has been split in two, with an explanation in the ADR accompanying this commit: the global and the component CSS.

The *global* CSS defines all of the variables that control the look and feel of the component, along with their defaults.

The *component* CSS reads those variables and applies them to the elements that make up its template. The topmost `:host` rule of the component CSS defines a layer that removes the per-component prefix, since it’s just repetitious noise, and duplicates the concrete default values as fallbacks. This allows the component to keep working well even when the global CSS is unavailable.

### Build system

The build system has been reworked into a more unix-flavored “do one thing and do it well.” Each script has a single role:

- Convert the Typescript into Javascript
- Convert the CSS and SCSS into component-level `.css.js` files
- Build the global CSS files

### Storybook

A special component, `ak-storybook-theme-provider`, has been provider as a decorator for *all* stories, so that we can control the theme. That decorator has been incorporated into the global decorator defined in `./storybook/preview`.

A theme decorator for storybook has been incorporated that watches the Storybook light/dark mode controller, and responds to changes in it by propogating those changes to the currently visible story.