This commit ports the Elements' build tools to Typescript
## What

All of the tools have been moved to a folder named `tools-src`, and the contents are compiled at the
start of the build process. The build has been organized around sequential and parallel processes,
but it still takes about 7 seconds with only one component in the works.

The ESLint file was updated to check the `tools-src` folder, and ignore the `tools` folder.

A new script, "copy-assets," replaces the awkward copy command we had earlier

We weren't using the `fontawesome` table, so out it goes. The CSS is fine as-is. The short names
file is managed by hand, anyway, since what gets shortnamed is a matter of taste. It's all patterns
anyway.

## Why

One of the tasks, the one that derives the CSS automatically from Patternfly, involves some fairly
complicated transformations; some are to derive the information from the CSS, others are to
transform that CSS into formats that make sense in a ShadowDOM context.  While working on it, using
just JavaScript, I was getting thoroughly lost and getting very little help from VSCode on how data
flowed through the system.

By giving them names, I was able to complete the work in very short order. During the conversion, I
decided to convert all the scripts. The result is more readable and reliable, as now the type
checker is validating that the steps being taken are correct even before the code is run.

(Due to its size, that converter is in a separate commit.)