This commit revises the WCCSS transpiler, using the regular expression and merge algorithms recommended by the Patternfly 5 React team.
## What

- `./tools-src/lib/generateTokens.mts`: This file is an extensive reworking of the Patternfly 5
  React team's `generateTokens.js`; I've ported it to Typescript and broken out a lot of
  sub-functions to help visualize what's going on within. This code reads the global Patternfly CSS
  and the CSS and SCSS for every component in Patternfly, and converts them to a tree of:

  ```
    component:
      selector:
        property:
          property: string
          value: string
          values: [string]
  ```

  In this scheme, the `values` field is the chain of definitions from where it appears in the
  component all the way up, through the CSS Custom Properties and then through the Sass variables,
  to the concrete definition.

  This allows us to pick a component and extract the information needed to populate the `:root`
  object with the component's declarations, and where those definitions come from, and populate the
  `:host` with a mapping of those declarations to a short-hand name as the property, and both the
  global BEMP value on the right; and if the global value is a CSS Custom Property, wrap it in a
  `var()` expression, providing the default value to ensure the component always has access to its
  default look and feel, even if the global CSS `:root` entries are missing.

- `./cludeRuleParser.mts`: This (short) file contains a minimal parser for understanding the format
  for rules that describe how and when to include a Patternfly CSS rule in a web component's CSS.
  Since the format is any combination of `((string | regex),)*\s*(string|regex)`, a small parser was
  the easiest way to go.

- `./tools/build-wccss-sheets.mts`: The script that uses the above to generate the CSS for a
  component. A YAML file describes the selectors that the component will use internally and, for
  each of those selectors, what from corresponding Patternfly React's CSS we will be importing. It
  performs transformations to build both a `component.css` file with the `:host` entries, and a
  `component.root.css` file with the `:root` entries, as described in the section on
  `generateTokens`.

These are both beastly big functions-- parsers and compilers often are, and the complexity of
Patternfly doesn't make this at all easy.  I spent a day just trying to find the right names,
patterns, and data flows through the system so that someone reading this code could easily
understand it.  But they're still big and they still take a lot of time to sink in.

# Why

The idea of automating the conversion of Patternfly's CSS to a Web Component CSS (WCCSS) format has
been around for awhile. The original WCCSS format is, well, nobody wants to learn a new DSL, even if
most of the DSL's logic has been preserved. The Patternfly React people supplied us with the
necessary and battle-hardened Regex's needed to extract that information, as well as some (somewhat
fragile) logic for backtracking and finding the concrete values.

Compressing this business down to a single script and a library makes it possible to keep this
in-house, as part of our development and build processes.