bug: fix poor formatting during the build process
## What

This commit adds `prettier.format` to the toolchain for CSS and SCSS files that are generated during
the build process by WCCSS.

## Why

These files are generated using the ReWork CSS Parser/Stringifier and are committed to the
repository. The stringifier's output does not conform with our chosen `prettier` formatting, which
triggers the `precommit` hook to complain. This fix ensures that any complaints by the `precommit`
hook are strictly caused by human error. :-)