element: ak-timeout
This commit adds the `timeout` component to our library, with an “elapsed time” indicator.

It also includes an event-monitoring decorator that changes in response to changes to media-queries. While in this case we use it to track `prefers-reduced-motion` (which restricts the `elapsed` time interval to a minimum of 60
seconds, since anything less than that is highly animated), it can be used to track any number of CSS queries, such as `display-mode`, `min-width`, and even `device-posture`.

`ak-timeout` uses a highly simplified event state mananger for updating the elapsed time. This is a crude metric, with a minimal resolution of 1000 milliseconds, so I avoided advanced techniques like animation frame. Using
`setTimeout()` with a “set the next timeout event” instead of `setInterval()` is generally preferred; the precision is lower, but there’s a heck of a lot less machinery, and stopping it is trivially simpler.

\# What

\# Why

\# How

\# Designs

\# Test Steps

\# Other Notes