infrastructure: husky & lint-staged
This commit introduces husky, lint-staged, and stylelint to the toolkit. On any commit attempt, husky will automatically run lint-staged via the `./.husky/pre-commit` script. `lint-staged` will then read the instructions is
package.json and run the linters listed in order in the `"lint-staged":` block, which right now are TSC types, ESlint `--fix`, and Prettier `--write`.

Stylelint is *very* picky, and since our CSS is machine-generated, this is a huge win; it already uncovered two issues in ak-button: although the test said that hover worked correctly on ak-button, it wasn’t because a regexp was
incorrect, and the list of CSS custom properties being sent to the final CSS was duplicated because of the use of the deprecated `:host` rule. Running `stylelint --fix` *after* `prettier --write` also allows us to enforce certain
rules with respect to spacing and ordering, although one *big* ordering issues cannot be resolved, which is why `.stylelintignore` has several vendored files; we’ll have to go through them later for fixes.

Two stylelint rules had to be modified to accept Patternfly’s CSS Custom Property naming scheme. It’s a valid scheme, it’s just… not what most people use. But it enables BEMP naming, which is the point.

Eslint now checks the WCC files for dumb Yaml errors. It can’t catch WCC syntax errors (yet).

\# What

\# Why

\# How

\# Designs

\# Test Steps

\# Other Notes