infrastructure: enforce `compat: baseline widely available` to our linting
This commit adds `browserslist` to our linting, with configurations added to both ESLint (for Javascript) and Stylelint (for CSS) to ensure that we don’t use features that, yeah, we would really *like* to use but can’t because not enough users have browsers that support it.

That said, a couple of progressive enhancements have been marked as “do not delete” in our CSS reset that we use inside components that need it. Stylelint was all, like, “These aren’t supported in Safari or Firefox.” Yeah, we know.

I have also added to Stylelint the “do not automatically convert to short-hand” any rules written in out in longhand. Patternfly leaves those in place to enable customization of some subtle “lighting” effects with drop shadows, and I’d like to keep that.