element: ak-tooltip
# What

This commit revises some of the tooltip handling code, mostly because it looked *terrible* in dark mode.

1.  Remove the `floating-ui` controls of the arrow. Instead, revise the CSS to use the `floating-ui` position names, and map them to the Patternfly Arrow positions.

    This lets us use the Patternfly-5 tooltip look and feel in its entirety, without having to try and do the math ourselves for where to place the arrow.

2.  Handle some subtle bugs in the CSS parsing that controls the Tooltip.

3.  Provide the Patternfly 5 “offset from anchor” as a CSS variable.

    - Required a whole new utility library to parse, interpret, and return the pixel size for a variety of CSS. Right now we handle Rem, Em, and Ch. (Well, and Px, of course).

4.  Provide the Dark Mode CSS, which, interestingly enough, exists only on `:root`. No changes to `:host` were necessary. (See Note 1!)

5.  Begins the revision of the Patternfly 5 globals to support dark mode as we understand it.

6.  Updates the test harness to the latest and greatest Chrome

There seems to be a little churn as other components are brought up-to-date with Stylelint’s clean-up pass, which is why `ak-button.css` got a few changes.

# Notes

**Note 1:**

This is a tactical rather than a strategic choice. By not having every dark-mode-aware element checking either the `<html>` tag or listening for its own thematic settings, we “admit” that the default look of the component is its light-mode, and that without the global CSS it will not respond to dark-mode settings. With the global CSS, of course, everything works fine.

It’s also important to note that the two Dark Mode global variables used by this element do not exist at all in light mode. This leads me to think that we need better names for these variables than the ones we picked up from Patternfly 5.